Implements exception handling concept in diff impl methods.

+++ demo/src/main/java/Genboot/demo/services/CloudVendorServicelmpl.java
@@ -1,8 +1,9 @@
 package Genboot.demo.services;
 
+import Genboot.demo.exception.CloudVendorAlreadyExistsException;
+import Genboot.demo.exception.CloudVendorNotFoundException;
 import Genboot.demo.model.CloudVendor;
 import Genboot.demo.repository.CloudVendorRepository;
-import jakarta.persistence.Id;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -16,15 +17,20 @@ private CloudVendorRepository cloudVendorRepository;
 
     @Override
     public String create(CloudVendor cloudVendor) {
+        if (cloudVendorRepository.findByVendorPhoneNumber(cloudVendor.getVendorPhoneNumber()).isPresent()) {
+            throw new CloudVendorAlreadyExistsException("Cloud vendor already exists");
+        }
         cloudVendorRepository.save(cloudVendor);
         return "Vendor created";
     }
 
+
     @Override
         public String update(Long Id, CloudVendor cloudVendor) {
-            cloudVendorRepository.findById(Id);
-        cloudVendorRepository.save(cloudVendor);
-        return "Vendor updated";
+            if (cloudVendorRepository.findById(Id).isEmpty())
+            throw  new CloudVendorNotFoundException("Requested cloud vendor does not exists.");
+        else{cloudVendorRepository.save(cloudVendor);
+        return "Vendor updated";}
     }
 
     @Override
@@ -33,15 +39,18 @@ private CloudVendorRepository cloudVendorRepository;
         return "Vendor Updated";
     }
 
-
     @Override
     public String delete(Long Id) {
+        if (cloudVendorRepository.findById(Id).isEmpty())
+            throw  new CloudVendorNotFoundException("Requested cloud vendor does not exists.");
         cloudVendorRepository.deleteById(Id);
         return "particular Vendor deleted";
     }
 
     @Override
     public CloudVendor get(Long Id) {
+        if(cloudVendorRepository.findById(Id).isEmpty())
+            throw  new CloudVendorNotFoundException("Requested cloud vendor does not exists.");
         return cloudVendorRepository.findById(Id).get();
 
     }
